requirements.txt
datasets
transformers==4.41.2
accelerate
pillow
torch
sentencepiece
torchvision
timm
einops